Implement editSquareRule()
This function is important because it alters the CSS stylesheet by removing the old rule for the class 'square'.
Then, it adds a new rule for the class 'square' while changing the flex-basis of each of the squares to reflect
the change in the number of squares per side of the grid.